location/requirements.txt
Flask
pandas
requests
Flask-RESTx
python-dotenv